requirements.txt
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
Babel
beautifulsoup4
bleach
cachetools
certifi
cffi
charset-normalizer
colorama
comm
contourpy
cycler
debugpy
decorator
defusedxml
exceptiongroup
executing
fastjsonschema
filelock
fonttools
fqdn
fsspec
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
idna
ipykernel
ipython
isoduration
jedi
Jinja2
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyter-events
jupyter-lsp
jupyter_server
jupyter-server-mathjax
jupyter_server_terminals
jupyterlab
jupyterlab_git
jupyterlab_pygments
jupyterlab_server
kiwisolver
MarkupSafe
matplotlib
matplotlib-inline
mistune
mpmath
nbclient
nbconvert
nbdime
nbformat
nest-asyncio
networkx
notebook_shim
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
overrides
packaging
pandas
pandocfilters
parso
pexpect
pillow
pip
platformdirs
plotly
prometheus_client
prompt-toolkit
proto-plus
protobuf
psutil
ptyprocess
pure-eval
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
Pygments
pyparsing
python-dateutil
python-json-logger
pytz
PyYAML
pyzmq
referencing
requests
rfc3339-validator
rfc3986-validator
rpds-py
rsa
scikit-learn
scipy
seaborn
Send2Trash
setuptools
six
smmap
sniffio
soupsieve
stack-data
sympy
tenacity
terminado
threadpoolctl
tinycss2
tomli
torch
tornado
tqdm
traitlets
triton
types-python-dateutil
typing_extensions
tzdata
uri-template
uritemplate
urllib3
wcwidth
webcolors
webencodings
websocket-client